refactor(pack): refactor xdsl ip portreset deconsolidation
