+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/deconsolidation/contract/contract.modal.html
