+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/ipv6/ipv6.html
